[clang] Introduce `SemaX86`
